wai-aria/state/basic.tentative.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>

<!--

These tests should remain solid and passing in any implementation that supports get_element_accessible_properties.

It uses a standard promise_test (rather than aria-utils.js) to reduce other dependencies.

If you're adding something you expect to fail in one or more implementations, you probably want a different file.

-->

<div id="checkboxCheckedTrue" role="checkbox" aria-checked="true"></div>
<div id="buttonPressedTrue" role="button" aria-pressed="true"></div>

<script>

promise_test(async t => {
  const acc = await test_driver.get_element_accessible_properties(document.getElementById("checkboxCheckedTrue"));
  assert_equals(acc.checked, "true");
}, "checkbox with aria-checked true");

promise_test(async t => {
  const acc = await test_driver.get_element_accessible_properties(document.getElementById("buttonPressedTrue"));
  assert_equals(acc.pressed, "true");
}, "button with aria-pressed true");

</script>
